+++ nym-api/src/support/http/mod.rs
@@ -25,25 +25,6 @@ use rocket_cors::{AllowedHeaders, AllowedOrigins, Cors};
 use rocket_okapi::mount_endpoints_and_merged_docs;
 use rocket_okapi::swagger_ui::make_swagger_ui;
 
-#[cfg(feature = "axum")]
-use {
-    crate::circulating_supply_api::handlers::circulating_supply_routes,
-    crate::network::handlers::{nym_network_routes, ContractVersionSchemaResponse},
-    crate::node_status_api::handlers::node_status_routes,
-    crate::nym_contract_cache::handlers::nym_contract_cache_routes,
-    crate::nym_nodes::handlers::nym_node_routes,
-    crate::nym_nodes::handlers_unstable::nym_node_routes_unstable,
-    crate::status,
-    crate::v2::AxumAppState,
-    axum::Router,
-    nym_api_requests::models,
-    nym_http_api_common::logging::logger,
-    tower_http::cors::CorsLayer,
-    utoipa::OpenApi,
-    utoipa_swagger_ui::SwaggerUi,
-    utoipauto::utoipauto,
-};
-
 pub(crate) mod helpers;
 pub(crate) mod openapi;
 
@@ -157,126 +138,4 @@ fn setup_rocket_cors() -> Result<Cors> {
     .to_cors()?;
 
     Ok(cors)
-}
-
-#[cfg(feature = "axum")]
-fn setup_cors() -> CorsLayer {
-    CorsLayer::new()
-        .allow_origin(tower_http::cors::Any)
-        .allow_methods([axum::http::Method::GET, axum::http::Method::POST])
-        .allow_headers(tower_http::cors::Any)
-        .allow_credentials(false)
-}
-
-// TODO once https://github.com/ProbablyClem/utoipauto/pull/38 is released:
-// include ",./nym-api/nym-api-requests/src from nym-api-requests" (and other packages mentioned below)
-// for automatic model discovery based on ToSchema / IntoParams implementation.
-// Then you can remove `components(schemas)` manual imports below
-
-#[cfg(feature = "axum")]
-#[utoipauto(paths = "./nym-api/src")]
-#[derive(OpenApi)]
-#[openapi(
-    info(title = "Nym API"),
-    tags(),
-    components(schemas(
-        models::CirculatingSupplyResponse,
-        models::CoinSchema,
-        nym_mixnet_contract_common::Interval,
-        nym_api_requests::models::GatewayStatusReportResponse,
-        nym_api_requests::models::GatewayUptimeHistoryResponse,
-        nym_api_requests::models::HistoricalUptimeResponse,
-        nym_api_requests::models::GatewayCoreStatusResponse,
-        nym_api_requests::models::GatewayUptimeResponse,
-        nym_api_requests::models::RewardEstimationResponse,
-        nym_api_requests::models::UptimeResponse,
-        nym_api_requests::models::ComputeRewardEstParam,
-        nym_api_requests::models::MixNodeBondAnnotated,
-        nym_api_requests::models::GatewayBondAnnotated,
-        nym_api_requests::models::MixnodeTestResultResponse,
-        nym_api_requests::models::StakeSaturationResponse,
-        nym_api_requests::models::InclusionProbabilityResponse,
-        nym_api_requests::models::AllInclusionProbabilitiesResponse,
-        nym_api_requests::models::InclusionProbability,
-        nym_api_requests::models::SelectionChance,
-        NetworkDetails,
-        nym_config::defaults::NymNetworkDetails,
-        nym_config::defaults::ChainDetails,
-        nym_config::defaults::DenomDetailsOwned,
-        nym_config::defaults::ValidatorDetails,
-        nym_config::defaults::NymContracts,
-        ContractVersionSchemaResponse,
-        crate::network::models::ContractInformation<ContractVersionSchemaResponse>,
-        nym_api_requests::models::ApiHealthResponse,
-        nym_api_requests::models::ApiStatus,
-        nym_bin_common::build_information::BinaryBuildInformationOwned,
-        nym_api_requests::models::SignerInformationResponse,
-        nym_api_requests::models::DescribedGateway,
-        nym_api_requests::models::MixNodeDetailsSchema,
-        nym_mixnet_contract_common::Gateway,
-        nym_mixnet_contract_common::GatewayBond,
-        nym_api_requests::models::NymNodeDescription,
-        nym_api_requests::models::HostInformation,
-        nym_api_requests::models::HostKeys,
-        nym_node_requests::api::v1::node::models::AuxiliaryDetails,
-        nym_api_requests::models::NetworkRequesterDetails,
-        nym_api_requests::models::IpPacketRouterDetails,
-        nym_api_requests::models::AuthenticatorDetails,
-        nym_api_requests::models::WebSockets,
-        nym_api_requests::nym_nodes::NodeRole,
-        nym_api_requests::models::DescribedMixNode,
-        nym_api_requests::ecash::VerificationKeyResponse,
-        nym_api_requests::ecash::models::AggregatedExpirationDateSignatureResponse,
-        nym_api_requests::ecash::models::AggregatedCoinIndicesSignatureResponse,
-        nym_api_requests::ecash::models::EpochCredentialsResponse,
-        nym_api_requests::ecash::models::IssuedCredentialResponse,
-        nym_api_requests::ecash::models::IssuedTicketbookBody,
-        nym_api_requests::ecash::models::BlindedSignatureResponse,
-        nym_api_requests::ecash::models::PartialExpirationDateSignatureResponse,
-        nym_api_requests::ecash::models::PartialCoinIndicesSignatureResponse,
-        nym_api_requests::ecash::models::EcashTicketVerificationResponse,
-        nym_api_requests::ecash::models::EcashTicketVerificationRejection,
-        nym_api_requests::ecash::models::EcashBatchTicketRedemptionResponse,
-        nym_api_requests::ecash::models::SpentCredentialsResponse,
-        nym_api_requests::ecash::models::IssuedCredentialsResponse,
-        nym_api_requests::nym_nodes::SkimmedNode,
-        nym_api_requests::nym_nodes::BasicEntryInformation,
-        nym_api_requests::nym_nodes::SemiSkimmedNode,
-        nym_api_requests::nym_nodes::NodeRoleQueryParam,
-    ))
-)]
-struct ApiDoc;
-
-#[cfg(feature = "axum")]
-pub(crate) async fn setup_routes(network_monitor: bool) -> anyhow::Result<Router<AxumAppState>> {
-    let router = Router::new()
-        // https://docs.rs/tower-http/0.1.1/tower_http/trace/index.html
-        // TODO rocket use tracing instead of env_logger
-        // https://github.com/tokio-rs/axum/blob/main/examples/tracing-aka-logging/src/main.rs
-        // .layer(
-        //     TraceLayer::new_for_http()
-        //         .make_span_with(DefaultMakeSpan::new().include_headers(true))
-        //         .on_request(DefaultOnRequest::new())
-        //         .on_response(DefaultOnResponse::new().latency_unit(tower_http::LatencyUnit::Micros)),
-        // )
-        // .route("/", axum::routing::get(|| async {axum::response::Redirect::permanent("/swagger")}))
-        // .route("/swagger", axum::routing::get(hello))
-        .merge(SwaggerUi::new("/swagger").url("/api-docs/openapi.json", ApiDoc::openapi()))
-        .nest(
-            "/v1",
-            Router::new()
-                .nest("/circulating-supply", circulating_supply_routes())
-                .merge(nym_contract_cache_routes())
-                .nest("/status", node_status_routes(network_monitor))
-                .nest("/network", nym_network_routes())
-                .nest("/api-status", status::handlers::api_status_routes())
-                .merge(nym_node_routes())
-                .nest("/unstable/nym-nodes", nym_node_routes_unstable()), // CORS layer needs to be "outside" of routes
-        )
-        .layer(setup_cors());
-
-    // needs to be after routes we want to log
-    let router = router.layer(axum::middleware::from_fn(logger));
-
-    Ok(router)
 }
